Fixes #WK Rant
"And where is our logo? I just want to say that your logo is being displayed on the site and I think we deserve to have ours on there too. And our, shouldn't it be in the header as well? We are mentioned on the site on multiple occassions. It may be confussing if we're not in the header as well"
...